SoloQueueIngest/requirements.txt
python-dotenv
cassiopeia
pymongo
black